+++ crates/teloxide-core/src/requests/requester.rs
@@ -968,6 +968,15 @@ pub trait Requester {
     where
         W: Into<String>;
 
+    type SavePreparedInlineMessage: Request<Payload = SavePreparedInlineMessage, Err = Self::Err>;
+
+    /// For Telegram documentation see [`SavePreparedInlineMessage`].
+    fn save_prepared_inline_message(
+        &self,
+        user_id: UserId,
+        result: InlineQueryResult,
+    ) -> Self::SavePreparedInlineMessage;
+
     type EditMessageText: Request<Payload = EditMessageText, Err = Self::Err>;
 
     /// For Telegram documentation see [`EditMessageText`].
@@ -1477,6 +1486,7 @@ macro_rules! forward_all {
             delete_my_commands,
             answer_inline_query,
             answer_web_app_query,
+            save_prepared_inline_message,
             edit_message_text,
             edit_message_text_inline,
             edit_message_caption,